requirements.txt
asgiref==3.5.2
Django==4.1
django-admin-sortable2==2.1.2
django-next-prev==1.1.0
django-tinymce==3.4.0
Pillow==9.2.0
sqlparse==0.4.2
